backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Pydantic
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Task queue
celery==5.4.0
redis==5.0.4
flower==2.0.1

# Scraping
playwright==1.44.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
fake-useragent==1.5.1
apify-client==1.7.1

# Outreach
twilio==9.1.0
sendgrid==6.11.0

# CRM
hubspot-api-client==9.0.0

# AI
anthropic==0.26.0

# Google Drive
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
phonenumbers==8.13.37
pendulum==3.0.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0
